backend/requirements.txt
# Trivector.ai Backend - WebSocket Collaboration Server
# Python 3.11+

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# WebSocket Support
websockets==12.0
python-socketio==5.11.0

# Authentication & Security
pyjwt[crypto]==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database (optional - for session persistence)
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis (for session state)
redis==5.0.1
aioredis==2.0.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# Monitoring & Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
